Proyecto Final - Jorge Monzo/requirements.txt
mysql-connector-python 
bcrypt